requirements.txt
Django>=4.2,<5.0
redis>=4.0.0
pandas>=1.5.0
openpyxl>=3.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0

# Production deployment
gunicorn>=21.2.0
whitenoise>=6.5.0

# Development and testing (optional)
coverage>=7.2.0
flake8>=6.0.0
black>=23.0.0

# Security
django-environ>=0.10.0